norminetted echo
