+++ minishell.h
@@ -97,18 +97,27 @@ typedef struct s_token
 	struct s_token		*next;
 }						t_token;
 
+/* ** BUILT IN FUNCS ** */
+/* ft_echo */
+int						ft_echo(char **argv, char ***envp_ptr);
+char					*get_variable_value(int index, char **envp_ptr);
+char					*modify_variable(char *str);
+char					*detect_varaible_name(char *argv);
+/* ft_cd */
+int						ft_cd(char **argv, char **envp);
+/* ft_pwd */
+int						ft_pwd(void);
+/* ft_export */
+int						ft_export(char **argv, char ***envp_ptr);
+// int						ft_unset(char **argv, char **envp);
+// int						ft_env(char **envp);
+
 void					save_terminal_state(struct termios *original_state);
 void					restore_terminal_state(struct termios *original_state);
 int						ft_strcmp(const char *s1, const char *s2);
 // void					update_env_var(char ***envp_ptr, const char *var_name,
 // 							const char *value);
-int						ft_echo(char **argv, char ***envp_ptr);
-int						check_exists(char **argv, char **envp);
-int						ft_cd(char **argv, char **envp);
-int						ft_pwd(void);
-int						ft_export(char **argv, char ***envp_ptr);
-// int						ft_unset(char **argv, char **envp);
-// int						ft_env(char **envp);
+
 int						ft_exit_builtin(char **argv, t_token *token,
 							char *line);
