+++ srcs/bltn_funcs/echo/ft_echo.c
@@ -12,33 +12,6 @@
 
 #include "../../../minishell.h"
 
-char	*detect_varaible_name(char *argv)
-{
-	int		i;
-	char	**strs;
-	char	*varaible_name;
-
-	strs = ft_split(argv, ' ');
-	varaible_name = NULL;
-	if (!strs)
-		return (NULL);
-	i = 0;
-	while (strs[i] != NULL)
-	{
-		if (strs[i][0] == '$')
-		{
-			varaible_name = ft_strdup(strs[i]);
-			free_tokens(strs);
-			if (!varaible_name)
-				return (NULL);
-			return (varaible_name);
-		}
-		i++;
-	}
-	free_tokens(strs);
-	return (NULL);
-}
-
 char	*process_arguments(char *str)
 {
 	char	*processed_str;
@@ -65,28 +38,6 @@ char	*process_arguments(char *str)
 	processed_str[j] = '\0';
 	return (processed_str);
 }
-char	*modify_variable(char *str)
-{
-	int		i;
-	int		j;
-	char	*modified_str;
-
-	if ((int)ft_strlen(str) < 0)
-		return (NULL);
-	i = 1;
-	j = 0;
-	modified_str = malloc(sizeof(char) * ft_strlen(str));
-	if (!modified_str)
-		return (NULL);
-	while (str[i] != '\0')
-	{
-		modified_str[j] = str[i];
-		i++;
-		j++;
-	}
-	modified_str[j] = '\0';
-	return (modified_str);
-}
 
 int	check_on_evnp(char *variable_name, char **envp)
 {
@@ -105,7 +56,7 @@ int	check_on_evnp(char *variable_name, char **envp)
 	while (envp[i] != NULL)
 	{
 		if (ft_strncmp(variable_name, envp[i], len) == 0 && (envp[i][len] == '='
-				|| envp[i][len] == '\0'))
+			|| envp[i][len] == '\0'))
 		{
 			free(variable_name);
 			return (i);
@@ -114,28 +65,6 @@ int	check_on_evnp(char *variable_name, char **envp)
 	}
 	free(variable_name);
 	return (-1);
-}
-char	*get_variable_value(int index, char **envp_ptr)
-{
-	char	*value;
-	int		j;
-	int		i;
-
-	i = 0;
-	while (envp_ptr[index][i] != '=')
-		i++;
-	i++;
-	j = (int)ft_strlen(envp_ptr[i]) - i;
-	value = malloc(sizeof(char) * j + 1);
-	j = 0;
-	while (envp_ptr[index][i] != '\0')
-	{
-		value[j] = envp_ptr[index][i];
-		j++;
-		i++;
-	}
-	value[j] = '\0';
-	return (value);
 }
 
 char	*check_variables(char *str, char ***envp_ptr)
